Refactored TranslationProvider to RootLayout
