+++ app/[locale]/(dashboard)/dashboard/save/[action]/[[...id]]/page.tsx
@@ -1,9 +1,9 @@
+import { useTranslation } from "react-i18next"
+
 import { actions } from "@/lib/Constants"
 import { DashboardHeader } from "@/components/Dashboard/dashboard-header"
 import { DashboardShell } from "@/components/Dashboard/dashboard-shell"
 import { EncryptionForm } from "@/components/Dashboard/encrypt/encrypt-form"
-import TranslationsProvider from "@/components/translations-provider"
-import initTranslations from "@/app/i18n"
 
 interface PageProps {
   params: {
@@ -13,27 +13,19 @@ interface PageProps {
   }
 }
 
-const i18nNamespaces = ["home", "common", "dashboard", "encrypt_form"]
-
 export default async function DashboardPage({
-  params: { action, id, locale },
+  params: { action, id },
 }: PageProps) {
-  const { t, resources } = await initTranslations(locale, i18nNamespaces)
+  const { t } = useTranslation()
   return (
-    <TranslationsProvider
-      resources={resources}
-      locale={locale}
-      namespaces={i18nNamespaces}
-    >
-      <DashboardShell>
-        <DashboardHeader
-          heading={t("dashboard:heading")}
-          text={t("dashboard:text")}
-        />
-        <div className="grid gap-8">
-          <EncryptionForm action={action as actions} id={id?.[0] ?? ""} />
-        </div>
-      </DashboardShell>
-    </TranslationsProvider>
+    <DashboardShell>
+      <DashboardHeader
+        heading={t("dashboard:heading")}
+        text={t("dashboard:text")}
+      />
+      <div className="grid gap-8">
+        <EncryptionForm action={action as actions} id={id?.[0] ?? ""} />
+      </div>
+    </DashboardShell>
   )
 }
