+++ components/translations-provider.tsx
@@ -20,7 +20,7 @@ export default function TranslationsProvider({
 }: TranslationsProviderProps) {
   const i18n = createInstance()
 
-  initTranslations(locale, namespaces, resources)
+  initTranslations(locale, namespaces, i18n, resources)
 
   return <I18nextProvider i18n={i18n}>{children}</I18nextProvider>
 }
